dichotomy-runner-api/src/main/java/com/farao_community/farao/dichotomy_runner/api/exception/DichotomyInternalException.java
/*
 * Copyright (c) 2021, RTE (http://www.rte-france.com)
 * This Source Code Form is subject to the terms of the Mozilla Public
 * License, v. 2.0. If a copy of the MPL was not distributed with this
 * file, You can obtain one at http://mozilla.org/MPL/2.0/.
 */
package com.farao_community.farao.dichotomy_runner.api.exception;

/**
 * @author Sebastien Murgey {@literal <sebastien.murgey at rte-france.com>}
 */
public class DichotomyInternalException extends AbstractDichotomyException {
    private static final int STATUS = 500;
    private static final String CODE = "500-InternalException";

    public DichotomyInternalException(String message) {
        super(message);
    }

    public DichotomyInternalException(String message, Throwable throwable) {
        super(message, throwable);
    }

    @Override
    public int getStatus() {
        return STATUS;
    }

    @Override
    public String getCode() {
        return CODE;
    }
}
